backend/requirements.txt
# Prefer Python 3.12 or 3.13. Python 3.14 may lack prebuilt wheels (e.g. pydantic-core) on Windows.
fastapi>=0.115.6,<0.116
uvicorn[standard]>=0.32.0,<0.33
python-dotenv>=1.0.1
pydantic>=2.11.0,<3
pydantic-settings>=2.6.0,<3
email-validator>=2.2.0

sqlalchemy[asyncio]>=2.0.36,<2.1
psycopg[binary,pool]>=3.2.10,<3.4
alembic>=1.14.0,<2
pgvector>=0.3.6

redis>=5.0.0,<6

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9

httpx>=0.27.0
sentry-sdk[fastapi]>=2.14.0
google-generativeai>=0.8.3
scipy>=1.14.1
PyPDF2>=3.0.1
pymupdf>=1.24.0
faiss-cpu>=1.9.0
numpy>=1.26.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
motor>=3.6.0
